Fix C interop portability and cross-build DLL/.so support

+++ array_urandomb.c
@@ -21,7 +21,7 @@
 
 void
 gmpmee_array_urandomb(mpz_t *rop, size_t len, gmp_randstate_t state,
-		   unsigned long int n) {
+     unsigned long long n) {
   size_t i;
 
   for (i = 0; i < len; i++)
